Add files via upload
Hey, 

I reorganized some of the  "Preparing the Data' section so it flowed a little better. We had some problems with headings up there. 

I started an analysis of my variables, its coming together nicely. I need to write up my description, but I know where I am going now. 

There is something weird with the Gini Index. It runs from 1800 through 2040, so it is longer than other variables. I am going to truncate this to the same range tomorrow, so we might see a change to that column of the pairs chart. 

Tomorrow EOD I am going to have my variables wrapped up - both on their own and modeling.